Fix overflow bug

+++ frakl/utils.h
@@ -16,9 +16,9 @@ static inline uintptr_t round_to_next_power_of_two(uintptr_t x) {
   if (!x) {
     return 0;
   }
-  const unsigned max_bit_set = sizeof(uintptr_t) * 8 - __round_clz(x) - 1;
+  const uintptr_t max_bit_set = sizeof(uintptr_t) * 8 - __round_clz(x) - 1;
   assert(max_bit_set < sizeof(uintptr_t) * 8);
-  const uintptr_t max_bit_value = 1 << max_bit_set;
+  const uintptr_t max_bit_value = (uintptr_t)1 << max_bit_set;
   if (max_bit_value == x) {
     return max_bit_value;
   } else {
